Fixes #1318 Save/load neighborhood
