+++ src/MoBi.Presentation/MenusAndBars/ContextMenus/ContextMenuFactoryForNeighborhoods.cs
@@ -1,6 +1,8 @@
-﻿using MoBi.Core.Domain.Model;
+﻿using MoBi.Assets;
+using MoBi.Core.Domain.Model;
 using MoBi.Presentation.DTO;
 using MoBi.Presentation.Presenter;
+using MoBi.Presentation.UICommand;
 using OSPSuite.Assets;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
@@ -9,7 +11,6 @@ using OSPSuite.Presentation.Core;
 using OSPSuite.Presentation.MenuAndBars;
 using OSPSuite.Presentation.Presenters;
 using OSPSuite.Presentation.Presenters.ContextMenus;
-using OSPSuite.Presentation.Repositories;
 using OSPSuite.Utility.Extensions;
 
 namespace MoBi.Presentation.MenusAndBars.ContextMenus
@@ -38,8 +39,8 @@ namespace MoBi.Presentation.MenusAndBars.ContextMenus
    public class ContextMenuForNeighborhoods : ContextMenuForContainerBase<IContainer>
    {
       public ContextMenuForNeighborhoods(
-         IMoBiContext context, 
-         IObjectTypeResolver objectTypeResolver, 
+         IMoBiContext context,
+         IObjectTypeResolver objectTypeResolver,
          OSPSuite.Utility.Container.IContainer container) : base(context, objectTypeResolver, container)
       {
       }
@@ -48,9 +49,16 @@ namespace MoBi.Presentation.MenusAndBars.ContextMenus
       {
          base.InitializeWith(dto, presenter);
          var neighborhoods = _context.Get<IContainer>(dto.Id);
-         //TODO add icon for neighborhood
-         _allMenuItems.Add(CreateAddNewChild<NeighborhoodBuilder>(neighborhoods).WithIcon(ApplicationIcons.ActiveEfflux).AsGroupStarter());
+         _allMenuItems.Add(createAddExistingChild(neighborhoods));
+         _allMenuItems.Add(CreateAddNewChild<NeighborhoodBuilder>(neighborhoods).WithIcon(ApplicationIcons.Neighborhood).AsGroupStarter());
          return this;
+      }
+
+      private IMenuBarItem createAddExistingChild(IContainer neighborhoodsContainer)
+      {
+         return CreateMenuButton.WithCaption(AppConstants.MenuNames.AddExisting(ObjectTypes.Neighborhood))
+            .WithIcon(ApplicationIcons.PKMLLoad)
+            .WithCommandFor<AddExistingCommandFor<IContainer, NeighborhoodBuilder>, IContainer>(neighborhoodsContainer, _container);
       }
    }
 }
